compile dependencies from source, bump libfuse
[ #185264727 ]

add ninja and meson packages for running the compilation steps of
libfuse / fuse mounthelper. Meson is not  the latest available version
and currently cannot be updated because newer versions of meson require
a newer python3 version than available on the xenial stemcell. This should
not matter much as meson itself never leaves the compilation VM since it is
only required to execute the build steps of fuse / libfuse.

remove handling of deb files from from pre-start script. The installation now
only requires symlinking the shared libraries into the default lookup paths
of the system.