compile dependencies from source, bump libfuse (#188)
* compile dependencies from source, bump libfuse

[ #185264727 ]

add ninja and meson packages for running the compilation steps of
libfuse / fuse mounthelper. Meson is not  the latest available version
and currently cannot be updated because newer versions of meson require
a newer python3 version than available on the xenial stemcell. This should
not matter much as meson itself never leaves the compilation VM since it is
only required to execute the build steps of fuse / libfuse.

remove handling of deb files from from pre-start script. The installation now
only requires symlinking the shared libraries into the default lookup paths
of the system.

* upload-blobs

* add option to use stemcell line

we want to test compilation of the release against jammy and xenial. currently
the test manifest only allows one stemcell that is hardcoded.

* remove dpkg related tests

we started compiling mapfs dependencies from source. This makes
the dpkg related tests obsolete because no dpkgs are used in pre-start
scripts anymore.

* remove obsolete assets

* remove build of obsolete asset

* make bosh release test deployment name configurable

so these tests can run against the same env as another set of bosh-release tests.

* specify deployment name to deploy command

* symlink required binary fusermount

nfsv3driver will try to call /bin/fusermount and the current
install location is not on $PATH for nfsv3driver.

---------

Co-authored-by: Cryogenics CI Bot <mapbu-cryogenics@groups.vmware.com>