fix symlinks to shared objects (#191)
* fix symlinks to shared objects

the current state creates possibly broken symlinks. Instead of symlinking
all discovered .so files, check whether the discovered .so file is a symlink itself
and create a symlink to the target to avoid ending up in a situation where
the symlinks end up in a broken state because of relative paths used for creating the
original symlink at compile time.

* check if file is present before trying to create a link

at the same location. it will fail.