fix bug introduced in 964e4a7 (#214)
[ #186969827 ]

Starting with v1.2.53 the release is broken in some deploy scenarios.

what works:

- deploying a new version of the release on a fresh VM ( e.g. on stemcell upgrade, anything that would recreate the VM)
- deploying the release on an existing VM without changing the release version

what breaks:
- upgrading an existing VM that already ran the pre-start script (of affected release version)
  without changing the stemcell

There are two issues with the current state of the script, the first one prevents
because of a missing `[` in the check for the existence of the paths for shared libs for lib-fuse.
But once that `[` is present, we will fail again because of the previously created symlinks at that location.

When an existing VM tries to upgrade this release, and run the new version of the pre-start script again,
it will fail with:

```
ln: failed to create symbolic link '/lib/libfuse3.so.3': File exists
```

In a deployment where the only difference is a change of this releases version, bosh will not recreate the vm

but the actual path of /var/vcap/packages/mapfs will change because that itself is a symlink.

In that scenario (upgrade a VM that has mapfs-release version v1.2.53) to a newer version ( e.g. a dev release still containing this bug)

- `[[ ! -f /lib/$LINK_NAME ]];` will return true. `-f` checks for the existence of a real file. Even a real file behind a symlink. since the VM did not get recreated, `/lib/libfuse3.so.3` exists, but it points to
`/var/vcap/data/<previous-package-version-shasum>/mapfs-fuse/lib/....` because the existing version of `/lib/libfuse3.so.3` was created using the result  `readlink -f /var/vcap/packages/mapfs-fuse/lib/.../libfuse3.so.3`.

To illustrate this:

ln -s /test/non-existing-target symlink-to-nonexisting
[[ -f ./symlink-to-nonexisting ]] => 1
touch /test/existing-target
ln -s /test/existing-target symlink-to-existing
[[ -f ./symlink-to-existing ]] => 0

it will then try to update the symlink and that will fail with:

```
ln: failed to create symbolic link '/lib/libfuse3.so.3': File exists
```

To enforce recreating the symlinks to the appropriate location, update the ln command to use `-fs` and remove the checks
because they are not necessary.

The 2nd. Issue present in this script is that it was able to produce

```
/var/vcap/jobs/mapfs/bin/pre-start: line 29: [: missing `]'
```
without exiting.

Yeah.. man sh:
```
           -e errexit       If not interactive, exit immediately if any untested command fails.  The exit status of a command is
                            considered to be explicitly tested if the command is used to control an if, elif, while, or until;
                            or if the command is the left hand operand of an “&&” or “||” operator.
```

and `[` is actually
`/usr/bin/[`

so what we're running is equivalent to:

```
if /bin/[ test ]]; then echo 'true'; else echo 'false'; fi
```

will output:

```
bash: /bin/[: No such file or directory
false
```

we're not calling the builtin `[[` we're calling `/usr/bin/[` and that is protected by the `the command is used to control an if`

it's weird but it's true:

```
bash  test
test: line 6: [: missing `]'
I'm somehow okay here
test: line 12: conditional binary operator expected
compute/5eab6de6-ffcd-4866-ab74-f1dbe553a0cf:/var/vcap/bosh_ssh/bosh_a41674b79dd34d3# cat test

set -eo pipefail

if [ true ]]; then
  echo "won't be printed"
else
  echo "I'm somehow okay here"
fi

if [[ true ]; then
  echo "won't be printed"
else
  echo "won't be printed either"
fi
echo "won't be printed finally"
```

Co-authored-by: Gareth Smith <sgareth@vmware.com>